requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
black==22.1.0
    # via -r requirements.in
click==8.0.4
    # via black
cloudpickle==2.0.0
    # via dask
colorama==0.4.4
    # via click
dask[array]==2022.2.1
    # via pycromanager
docstring-inheritance==1.0.0
    # via pycromanager
fsspec==2022.2.0
    # via dask
imageio==2.16.1
    # via scikit-image
locket==0.2.1
    # via partd
mypy-extensions==0.4.3
    # via black
networkx==2.7.1
    # via scikit-image
nidaqmx==0.6.1
    # via -r requirements.in
numpy==1.22.3
    # via
    #   dask
    #   imageio
    #   nidaqmx
    #   pycromanager
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
packaging==21.3
    # via
    #   dask
    #   scikit-image
partd==1.2.0
    # via dask
pathspec==0.9.0
    # via black
pillow==9.0.1
    # via
    #   -r requirements.in
    #   imageio
    #   scikit-image
platformdirs==2.5.1
    # via black
pycromanager==0.17.0
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
pypylon==1.8.0
    # via -r requirements.in
pyserial==3.5
    # via -r requirements.in
pywavelets==1.3.0
    # via scikit-image
pyyaml==6.0
    # via dask
pyzmq==22.3.0
    # via pycromanager
scikit-image==0.19.2
    # via -r requirements.in
scipy==1.8.0
    # via scikit-image
six==1.16.0
    # via nidaqmx
tifffile==2022.3.25
    # via
    #   -r requirements.in
    #   scikit-image
tomli==2.0.1
    # via black
toolz==0.11.2
    # via
    #   dask
    #   partd
typing-extensions==4.1.1
    # via black
